requirements.txt
Flask
Flask-Cors
pdfminer.six
python-docx
youtube-search-python
